requirements.txt
wheel
pytest
black
discord.py>=1.4.0
